[swift_libs] Fix exception with FutureTime.get_seconds
Some explanation may be required for why this happened:

The keys in `TIME_QUANTITIES` were LazyStrings once, and I just happened to revert that change before properly committing it,
and forgot to remove the extra call, which means that instead of calling a LazyString object, it was attempting to call a str.
Which of course, isn't possible.